Make the query read/write AttributeBuffer a common struct. #832
tiledb::sm namespace now contains the AttributeBuffer that is used by
tiledb::sm::Reader and tiledb::sm::Writer.
